+++ backend/.env
@@ -1,3 +1,2 @@
-
-DB_CONNECTION='mongodb://admin:palendar2019@ds123556.mlab.com:23556/palendar_db'
 PORT=8080
+MONGO_CONNECTION="mongodb://palendarql:palendar2019@ds123556.mlab.com:23556/palendar_db"
