+++ backend/app/graphql/schemas/User.ts
@@ -1,0 +1,8 @@
+const userDefs = `
+    type User {
+        _id: Int!
+        
+    }
+
+
+`
