+++ backend/app/models/user.ts
@@ -1,0 +1,25 @@
+import mongoose from 'mongoose';
+
+const Schema = mongoose.Schema;
+
+const userSchema = new Schema({
+    firstName: {
+        type: String,
+        required: true
+    },
+    lastName: {
+        type: String,
+        required: true
+    },
+    email: {
+        type: String,
+        required: true
+    },
+    password: {
+        type: String,
+        required: true
+    }
+});
+
+module.exports = mongoose.model('User',userSchema);
+
