--- backend/models/Events.ts
@@ -1,1 +1,0 @@
-import mongoose from 'mongoose';
