+++ backend/package.json
@@ -1,22 +1,37 @@
 {
   "name": "backend",
   "version": "1.0.0",
-  "description": "palendars backend",
+  "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
+    "tsc": "tsc",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^6.2.0",
+    "errorhandler": "^1.5.0",
+    "es6-shim": "^0.35.4",
     "express": "^4.16.4",
     "express-graphql": "^0.7.1",
     "graphql": "^14.1.1",
-    "jsonwebtoken": "^8.4.0",
-    "mongoose": "^5.4.7",
-    "nodemon": "^1.18.9"
+    "graphql-subscriptions": "^1.0.0",
+    "mongoose": "^5.4.9",
+    "typescript": "^3.3.1"
+  },
+  "devDependencies": {
+    "@types/graphql": "^14.0.5",
+    "@types/async": "^2.4.0",
+    "@types/cors": "^2.8.4",
+    "@types/dotenv": "^6.1.0",
+    "@types/errorhandler": "0.0.32",
+    "@types/es6-shim": "^0.31.39",
+    "@types/express": "^4.16.1",
+    "@types/express-graphql": "^0.6.2",
+    "@types/mongoose": "^5.3.12",
+    "@types/ws": "^6.0.1"
   }
 }
